requirements.txt
Flask==1.0.2
werkzeug==0.15.1
python-dotenv

# Extensions.
flask-debugtoolbar==0.10.1
Flask-Mail==0.9.1
Flask-Migrate
Flask-RESTful
Flask-Script
Flask-SQLAlchemy

# Application server for both development and production.
gunicorn==19.9.0
